requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
django-enumchoicefield==2.0.0
django-enumfields==2.1.1
django-extensions==3.1.5
django-model-utils==4.2.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-enum-field==0.9.3
graphviz==0.19.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
pygraphviz==1.6
pytz==2021.3
PyYAML==6.0
requests==2.27.1
simplejson==3.17.6
sqlparse==0.4.2
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
